Stat method for geom_fields

Prepares data before being handled by geom_fields()

StatFields

stat_fields(
  mapping = NULL,
  data = NULL,
  geom = "fields",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

Format

An object of class StatFields (inherits from StatSf, Stat, ggproto, gg) of length 3.

Arguments

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

Can be one of four things:

NULL: in that case data from the parent ggplot call is inherited.

data.frame: you need to assign the x and y aesthetics.

sf object: it should contain a geometry column with only POINT geometries.

stars object: it will be converted automatically to an sf object.

geom

The layer type for which the data is prepared. In this case "fields".

position

Position adjustment, either as a string naming the adjustment (e.g. "jitter" to use position_jitter), or the result of a call to a position adjustment function. Use the latter if you need to change the settings of the adjustment.

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Value

Returns a layer that can be further modified by geom_fields().

Author

Pepijn de Vries

Examples

stat_fields()
#> geom_fields: na.rm = FALSE
#> stat_fields: na.rm = FALSE
#> position_identity